+++ CHANGELOG.md
@@ -2,10 +2,12 @@
 
 ![DIIVE](images/logo_diive1_256px.png)
 
-## v0.71.7 | 24 Mar 2024
+## v0.72.0 | XX Mar 2024
 
 - Refactored code for class `HeatmapDateTime` (`diive.core.plotting.heatmap_datetime.HeatmapDateTime`)
 - Fixed bug in `HeatmapDateTime` where the last record of each day was not shown
+- (`diive.core.plotting.heatmap_base.HeatmapBase`)
+- (`diive.core.plotting.heatmap_datetime.HeatmapYearMonth`)
 
 ## v0.71.6 | 23 Mar 2024